pyproject.toml
[project]
name = "video-server"
version = "0.1.0"
description = "FastAPI video server with quote search and video streaming"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.125.0",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.38.0",
]

[project.scripts]
video-server = "app.main:main"

[tool.uv]
dev-dependencies = []
